@@ -6,22 +6,9 @@
  * building robust, powerful web applications using Vue and Laravel.
  */
 
-//const { default: Echo } = require('laravel-echo');
-
 require('./bootstrap');
 
 window.Vue = require('vue');
-
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
-
-const files = require.context('./', true, /\.vue$/i)
-files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -29,6 +16,20 @@ files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-const app = new Vue({
-    el: '#app',
+import { createApp } from 'vue'
+import router from './router'
+import store from './store'
+import App from './views/AppView'
+import NProgress from 'nprogress'
+
+NProgress.configure({ showSpinner: false });
+
+const app = createApp(App)
+
+store.dispatch('me').then(() => {
+    app.use(router);
+    app.use(store);
+
+    app.mount('#app');
 });
+
