@@ -1,17 +1,4 @@
-window._ = require('lodash');
 
-/**
- * We'll load jQuery and the Bootstrap jQuery plugin which provides support
- * for JavaScript based Bootstrap features such as modals and tabs. This
- * code may be modified to fit the specific needs of your application.
- */
-
-try {
-    window.Popper = require('popper.js').default;
-    window.$ = window.jQuery = require('jquery');
-
-    require('bootstrap');
-} catch (e) {}
 
 /**
  * We'll load the axios HTTP library which allows us to easily issue requests
@@ -19,10 +6,16 @@ try {
  * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-window.axios = require('axios');
+import Echo from 'laravel-echo';
+
+
+window._ = require('lodash');
+
+window.axios = require('axios').default;
 
 window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 window.axios.defaults.headers.common['Accept'] = 'application/json';
+window.axios.defaults.withCredentials = true;
 
 /**
  * Echo exposes an expressive API for subscribing to channels and listening
@@ -30,7 +23,7 @@ window.axios.defaults.headers.common['Accept'] = 'application/json';
  * allows your team to easily build robust real-time web applications.
  */
 
-import Echo from 'laravel-echo';
+
 
 window.Pusher = require('pusher-js');
 
