@@ -1,15 +1,18 @@
 <template>
     <img class="spinner" :src="link" alt="loading">
 </template>
-<script>
-    export default {
-        props: {
-            link: {
-                type: String,
-                required: true
-            }
+
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
+    props: {
+        link: {
+            type: String,
+            required: true
         }
     }
+})
 </script>
 
 <style scoped>
